Reorganize data sources, remove configurator
