+++ src/aslm/controller/sub_controllers/channels_tab_controller.py
@@ -129,8 +129,8 @@ class ChannelsTabController(GUI_Controller):
 
         Parameters
         ----------
-        config : object
-            ASLM_Configuration_Controller.  config.configuration = Configurator object.
+        config : multiprocesing.managers.DictProxy
+            Global configuration of the microscope
         """
         config = self.parent_controller.configuration_controller
