+++ src/aslm/model/devices/camera/camera_base.py
@@ -47,11 +47,11 @@ class CameraBase:
 
     Parameters
     ----------
-    microscope_name : str
-        Selects which camera to connect to the microscope.
-    device_connection: Object
-        Camera API object
-    configuration : Configurator
+   microscope_name : str
+        Name of microscope in configuration
+    device_connection : object
+        Hardware device to connect to
+    configuration : multiprocesing.managers.DictProxy
         Global configuration of the microscope
 
     """
