+++ src/aslm/model/devices/camera/camera_synthetic.py
@@ -74,12 +74,12 @@ class SyntheticCamera(CameraBase):
 
     Parameters
     ----------
-    camera_id : int
-        Selects which camera to connect to (0, 1, ...).
-    configuration : Configurator
+    microscope_name : str
+        Name of microscope in configuration
+    device_connection : object
+        Hardware device to connect to
+    configuration : multiprocesing.managers.DictProxy
         Global configuration of the microscope
-    experiment : Configurator
-        Experiment configuration of the microscope
 
     """
     def __init__(self, microscope_name, device_connection, configuration):
