+++ src/aslm/model/devices/shutter/laser_shutter_base.py
@@ -48,14 +48,12 @@ class ShutterBase:
 
     Attributes
     ----------
-    configuration : Configurator
+    microscope_name : str
+        Name of microscope in configuration
+    device_connection : object
+        Hardware device to connect to
+    configuration : multiprocesing.managers.DictProxy
         Global configuration of the microscope
-    experiment : Configurator
-        Experiment configuration of the microscope
-    shutter_right_state : bool
-        Right shutter state
-    shutter_left_state : bool
-        Left shutter state
     """
 
     def __init__(self, microscope_name, device_connection, configuration):
