+++ src/aslm/model/metadata_sources/bdv_metadata.py
