+++ src/aslm/model/metadata_sources/ome_tiff_metadata.py
