+++ test/model/devices/camera/test_camera_hamamatsu.py
@@ -38,8 +38,6 @@ from pathlib import Path
 
 # Local Imports
 from aslm.model.devices.camera.camera_hamamatsu import HamamatsuOrca
-from aslm.model.aslm_model_config import Configurator
-
 
 class TestHamamatsuOrca(unittest.TestCase):
     r"""Unit Test for HamamamatsuOrca Class"""
