+++ test/model/devices/shutter/test_laser_shutter_base.py
@@ -31,7 +31,6 @@ POSSIBILITY OF SUCH DAMAGE.
 """
 import unittest
 from pathlib import Path
-from aslm.model.aslm_model_config import Configurator
 from aslm.model.devices.shutter.laser_shutter_base import ShutterBase
 from aslm.model.dummy import DummyModel
