+++ test/model/model_features/data_sources/test_bdv_data_source.py
@@ -4,7 +4,7 @@ import numpy as np
 
 def test_bdv_write():
     from aslm.model.dummy import DummyModel
-    from aslm.model.model_features.data_sources.bdv_data_source import BigDataViewerDataSource
+    from aslm.model.data_sources.bdv_data_source import BigDataViewerDataSource
 
     # Set up model with a random number of z-steps to modulate the shape
     model = DummyModel()
