+++ test/model/model_features/test_aslm_image_writer.py
@@ -1,4 +1,4 @@
-from aslm.model.model_features.aslm_image_writer import ImageWriter
+from aslm.model.features.image_writer import ImageWriter
 from aslm.model.dummy import DummyModel
 import numpy as np
 import pytest
